feat(deploy): add GitHub Actions workflow for deploying the application
This commit adds a new file `.github/actions/deploy/action.yml` which contains the configuration for a GitHub Actions workflow. The workflow is triggered on push events and is responsible for deploying the application.

The workflow consists of the following steps:
- Checkout the repository
- Set up JDK 11 using the `actions/setup-java` action
- Build the application using Maven
- Publish the application to GitHub Packages Apache Maven repository

The `settings.xml` file is used for configuring the Maven deployment repository. The `GITHUB_TOKEN` environment variable is used for authentication when publishing to GitHub Packages.

This workflow enables automated deployment of the application whenever changes are pushed to the repository.